feat: move derive address helper function from contracts-sdk to misc
