+++ packages/contracts-sdk/src/index.ts
@@ -1,3 +1,2 @@
 export * from './lib/contracts-sdk';
-export * from './lib/addresses';
 export * from './lib/utils';
